+++ blocklets/ai-studio/api/src/templates/projects.ts
@@ -14,7 +14,6 @@ export const projectTemplates: {
       id: '363299428078977024',
       name: 'blank',
       description: 'Start a project from scratch.',
-      model: '',
       gitDefaultBranch: 'main',
       createdBy: wallet.address,
       updatedBy: wallet.address,